+++ src/Traits/Search.php
@@ -36,13 +36,10 @@ trait Search
         }
         
         // Preprocess the keys to search
-        $searchableKeys = collect($data->first() ?? [])->keys()->filter(function ($key) use ($data) {
-            // Include keys ending with "_search" or those without corresponding "_search" keys
-            if (str_ends_with($key, '_original')) {
-                return true;
-            }
-            $baseKey = preg_replace('/_search$/', '', $key);
-            return !array_key_exists($baseKey . '_original', $data->first());
+        $searchableKeys = $this->getFreshColumns()->filter(function ($column) {
+            return $column->isSearchable;
+        })->map(function ($column) {
+            return $column->key;
         });
 
         // Filter the collection